EES-5639 BE - Updating Controller tests
