+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Controllers/Api/ReleaseVersionsControllerTests.cs
@@ -42,26 +42,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
         private readonly Guid _publicationId = Guid.NewGuid();
 
         [Fact]
-        public async Task Create_Release_Returns_Ok()
-        {
-            var returnedViewModel = new ReleaseVersionViewModel();
-
-            var releaseVersionService = new Mock<IReleaseVersionService>(Strict);
-
-            releaseVersionService
-                .Setup(s => s.CreateRelease(It.IsAny<ReleaseCreateRequest>()))
-                .ReturnsAsync(returnedViewModel);
-
-            var controller = BuildController(releaseVersionService.Object);
-
-            // Call the method under test
-            var result = await controller.CreateRelease(new ReleaseCreateRequest(), _publicationId);
-            VerifyAllMocks(releaseVersionService);
-
-            result.AssertOkResult(returnedViewModel);
-        }
-
-        [Fact]
         public async Task UploadDataSet_Success()
         {
             var dataFile = MockFile("datafile.csv");
@@ -557,208 +537,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
 
     public abstract class ReleaseVersionsControllerIntegrationTests(TestApplicationFactory testApp) : IntegrationTestFixture(testApp)
     {
-        public class CreateReleaseTests(TestApplicationFactory testApp) : ReleaseVersionsControllerIntegrationTests(testApp)
-        {
-            [Theory]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "initial", "initial", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "Initial", "Initial", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, " initial", "initial", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "initial ", "initial", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "initial 2", "initial 2", "2020-21-initial-2")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "initial  2", "initial  2", "2020-21-initial-2")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "", null, "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, " ", null, "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "  ", null, "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, null, null, "2020-21")]
-            public async Task Success(
-                int year,
-                TimeIdentifier timePeriodCoverage,
-                string? label,
-                string? expectedLabel,
-                string expectedSlug)
-            {
-                Publication publication = DataFixture.DefaultPublication();
-
-                await TestApp.AddTestData<ContentDbContext>(
-                    context => context.Publications.Add(publication));
-
-                var response = await CreateRelease(
-                    publicationId: publication.Id,
-                    year: year,
-                    timePeriodCoverage: timePeriodCoverage,
-                    label: label);
-
-                var viewModel = response.AssertOk<ReleaseVersionViewModel>();
-
-                var contentDbContext = TestApp.GetDbContext<ContentDbContext>();
-
-                var updatedPublication = contentDbContext.Publications
-                    .Include(p => p.Releases)
-                    .ThenInclude(r => r.Versions)
-                    .Single(p => p.Id == publication.Id);
-
-                Assert.Equal(publication.Id, viewModel.PublicationId);
-                Assert.Equal(year, viewModel.Year);
-                Assert.Equal(timePeriodCoverage, viewModel.TimePeriodCoverage);
-                Assert.Equal(expectedSlug, viewModel.Slug);
-
-                var release = Assert.Single(updatedPublication.Releases);
-                Assert.Equal(publication.Id, release.PublicationId);
-                Assert.Equal(year, release.Year);
-                Assert.Equal(timePeriodCoverage, release.TimePeriodCoverage);
-                Assert.Equal(expectedLabel, release.Label);
-                Assert.Equal(expectedSlug, release.Slug);
-
-                var releaseVersion = Assert.Single(release.Versions);
-                Assert.Equal(publication.Id, releaseVersion.PublicationId);
-                Assert.Equal(release.Id, releaseVersion.ReleaseId);
-
-                var releaseSeriesItem = Assert.Single(updatedPublication.ReleaseSeries);
-                Assert.Equal(release.Id, releaseSeriesItem.ReleaseId);
-            }
-
-            [Fact]
-            public async Task PublicationNotFound()
-            {
-                var response = await CreateRelease(
-                    publicationId: Guid.NewGuid(),
-                    year: 2020,
-                    timePeriodCoverage: TimeIdentifier.AcademicYear,
-                    label: null);
-
-                response.AssertNotFound();
-            }
-
-            [Fact]
-            public async Task UserDoesNotHavePermission()
-            {
-                Publication publication = DataFixture.DefaultPublication();
-
-                await TestApp.AddTestData<ContentDbContext>(
-                    context => context.Publications.Add(publication));
-
-                var client = BuildApp(DataFixture.AuthenticatedUser()).CreateClient();
-
-                var response = await CreateRelease(
-                    publicationId: publication.Id,
-                    year: 2020,
-                    timePeriodCoverage: TimeIdentifier.AcademicYear,
-                    label: null,
-                    client: client);
-
-                response.AssertForbidden();
-            }
-
-            [Fact]
-            public async Task ReleaseTypeInvalid()
-            {
-                Publication publication = DataFixture.DefaultPublication();
-
-                await TestApp.AddTestData<ContentDbContext>(
-                    context => context.Publications.Add(publication));
-
-                var response = await CreateRelease(
-                    publicationId: publication.Id,
-                    year: 2020,
-                    timePeriodCoverage: TimeIdentifier.AcademicYear,
-                    label: null,
-                    type: ReleaseType.ExperimentalStatistics);
-
-                var validationProblem = response.AssertValidationProblem();
-
-                var error = Assert.Single(validationProblem.Errors);
-
-                Assert.Equal(ValidationErrorMessages.ReleaseTypeInvalid.ToString(), error.Code);
-            }
-
-            [Theory]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "initial", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "Initial", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, " initial ", "2020-21-initial")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, null, "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "", "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, " ", "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYear, "  ", "2020-21")]
-            [InlineData(2020, TimeIdentifier.AcademicYearQ1, "initial", "2020-21-q1-initial")]
-            public async Task SlugNotUniqueToPublication(
-                int year,
-                TimeIdentifier timePeriodCoverage,
-                string? label,
-                string existingReleaseSlug)
-            {
-                Publication publication = DataFixture.DefaultPublication()
-                    .WithReleases([
-                        DataFixture
-                            .DefaultRelease(publishedVersions: 1)
-                            .WithSlug(existingReleaseSlug)
-                        ]);
-
-                await TestApp.AddTestData<ContentDbContext>(
-                    context => context.Publications.Add(publication));
-
-                var response = await CreateRelease(
-                    publicationId: publication.Id,
-                    year: year,
-                    timePeriodCoverage: timePeriodCoverage,
-                    label: label);
-
-                var validationProblem = response.AssertValidationProblem();
-
-                var error = Assert.Single(validationProblem.Errors);
-
-                Assert.Equal(SlugNotUnique.ToString(), error.Code);
-            }
-
-            [Fact]
-            public async Task LabelOver50Characters()
-            {
-                Publication publication = DataFixture.DefaultPublication();
-
-                await TestApp.AddTestData<ContentDbContext>(
-                    context => context.Publications.Add(publication));
-
-                var response = await CreateRelease(
-                    publicationId: publication.Id,
-                    year: 2020,
-                    timePeriodCoverage: TimeIdentifier.AcademicYear,
-                    label: new string('a', 51));
-
-                var validationProblem = response.AssertValidationProblem();
-
-                var error = Assert.Single(validationProblem.Errors);
-
-                Assert.Equal($"The field {nameof(ReleaseCreateRequest.Label)} must be a string or array type with a maximum length of '50'.", error.Message);
-                Assert.Equal(nameof(ReleaseCreateRequest.Label), error.Path);
-            }
-
-            private WebApplicationFactory<TestStartup> BuildApp(
-                ClaimsPrincipal? user = null)
-            {
-                return TestApp.SetUser(user ?? DataFixture.BauUser());
-            }
-
-            private async Task<HttpResponseMessage> CreateRelease(
-                Guid publicationId,
-                int year,
-                TimeIdentifier timePeriodCoverage,
-                string? label = null,
-                ReleaseType? type = ReleaseType.OfficialStatistics,
-                HttpClient? client = null)
-            {
-                client ??= BuildApp().CreateClient();
-
-                var request = new
-                {
-                    Type = type,
-                    Year = year,
-                    TimePeriodCoverage = new { Value = timePeriodCoverage.GetEnumValue() },
-                    Label = label,
-                };
-
-                return await client.PostAsJsonAsync($"api/publications/{publicationId}/releases", request);
-            }
-        }
-
         public class UpdateReleaseTests(TestApplicationFactory testApp) : ReleaseVersionsControllerIntegrationTests(testApp)
         {
             [Fact]
@@ -1126,7 +904,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
                     PreReleaseAccessList = preReleaseAccessList
                 };
 
-                return await client.PutAsJsonAsync($"api/releases/{releaseVersionId}", request);
+                return await client.PatchAsJsonAsync($"api/releaseVersions/{releaseVersionId}", request);
             }
         }
     }
